Fix build and connectedCheck
