--- motor-control-module/src/androidTest/java/org/sagebionetworks/research/motor_control_module/ExampleInstrumentedTest.java
@@ -1,25 +1,0 @@
-package org.sagebionetworks.research.motor_control_module;
-
-import android.content.Context;
-import android.support.test.InstrumentationRegistry;
-import android.support.test.runner.AndroidJUnit4;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-import static org.junit.Assert.assertEquals;
-
-/**
- * Instrumented test, which will execute on an Android device.
- *
- * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
- */
-@RunWith(AndroidJUnit4.class)
-public class ExampleInstrumentedTest {
-    @Test
-    public void useAppContext() {
-        // Context of the app under test.
-        Context appContext = InstrumentationRegistry.getTargetContext();
-
-        assertEquals("org.sagebionetworks.research.motor_control_module.test", appContext.getPackageName());
-    }
-}
