--- motor-control-module/src/test/java/org/sagebionetworks/research/motor_control_module/ExampleUnitTest.java
@@ -1,17 +1,0 @@
-package org.sagebionetworks.research.motor_control_module;
-
-import org.junit.Test;
-
-import static org.junit.Assert.assertEquals;
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
- */
-public class ExampleUnitTest {
-    @Test
-    public void addition_isCorrect() {
-        assertEquals(4, 2 + 2);
-    }
-}
