+++ researchstack-mpower/build.gradle
@@ -12,6 +12,7 @@ android {
         versionName "1.0"
 
         consumerProguardFiles 'proguard-rules.pro'
+        multiDexEnabled true
 
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
     }
@@ -35,6 +36,7 @@ dependencies {
 
     // AndroidX
     implementation "com.android.support:appcompat-v7:$support_library_version"
+    implementation 'com.android.support:multidex:1.0.3'
 
     implementation 'com.google.dagger:dagger-android:2.16'
     implementation 'com.google.dagger:dagger-android-support:2.16'
