--- researchstack-mpower/src/test/java/org/sagebionetworks/research/mpower/researchstack/researchstack_mpower/ExampleUnitTest.java
@@ -1,17 +1,0 @@
-package com.example.researchstack_mpower;
-
-import static org.junit.Assert.assertEquals;
-
-import org.junit.Test;
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
- */
-public class ExampleUnitTest {
-    @Test
-    public void addition_isCorrect() {
-        assertEquals(4, 2 + 2);
-    }
-}
